triangle function
function gathering 4 values of 2 points, to make single line, then making from that line a triangle and returning the space of triangle. the space is checked if its the smallest or biggest. trojkat function contains a failsafe if the a value equals 0, which breaks the computing. if a = 0, it returns -1, that will be checked in main function. if the main function recognizes -1, the space value will be skipped.